+++ datanode/gateway/graphql/resolvers.go
@@ -416,29 +416,10 @@ func (r *myQueryResolver) Transfers(
 	return response.Transfers, nil
 }
 
-func (r *myQueryResolver) TransfersConnection(ctx context.Context, pubkey *string, direction TransferDirection,
+func (r *myQueryResolver) TransfersConnection(ctx context.Context, partyID *string, direction *TransferDirection,
 	pagination *v2.Pagination,
 ) (*v2.TransferConnection, error) {
-	var transferDirection v2.TransferDirection
-	switch direction {
-	case TransferDirectionFrom:
-		transferDirection = v2.TransferDirection_TRANSFER_DIRECTION_TRANSFER_FROM
-	case TransferDirectionTo:
-		transferDirection = v2.TransferDirection_TRANSFER_DIRECTION_TRANSFER_TO
-	case TransferDirectionToOrFrom:
-		transferDirection = v2.TransferDirection_TRANSFER_DIRECTION_TRANSFER_TO_OR_FROM
-	}
-
-	res, err := r.tradingDataClientV2.ListTransfers(ctx, &v2.ListTransfersRequest{
-		Pubkey:     pubkey,
-		Direction:  transferDirection,
-		Pagination: pagination,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return res.Transfers, nil
+	return r.r.transfersConnection(ctx, partyID, direction, pagination)
 }
 
 func (r *myQueryResolver) LastBlockHeight(ctx context.Context) (string, error) {
@@ -1192,6 +1173,15 @@ func (r *myPartyResolver) Rewards(
 	}
 	resp, err := r.tradingDataClient.GetRewards(ctx, req)
 	return resp.Rewards, err
+}
+
+func (r *myPartyResolver) TransfersConnection(
+	ctx context.Context,
+	party *types.Party,
+	direction *TransferDirection,
+	pagination *v2.Pagination,
+) (*v2.TransferConnection, error) {
+	return r.r.transfersConnection(ctx, &party.Id, direction, pagination)
 }
 
 func (r *myPartyResolver) RewardsConnection(ctx context.Context, party *types.Party, assetID *string, pagination *v2.Pagination) (*v2.RewardsConnection, error) {