+++ CHANGELOG.md
@@ -40,6 +40,9 @@ The init command now also generate the configuration for tendermint, the flags `
 - [5503](https://github.com/vegaprotocol/vega/issues/5503) - Introduce wallet API version 2 based on JSON-RPC with new authentication workflow
 - [5822](https://github.com/vegaprotocol/vega/issues/5822) - Rename `Graphql` enums
 - [5618](https://github.com/vegaprotocol/vega/issues/5618) - Add wallet JSON-RPC documentation
+- [5685](https://github.com/vegaprotocol/vega/issues/5685) - Migrated `apis` from `datanode v1` to `datanode v2`
+- [5761](https://github.com/vegaprotocol/vega/issues/5761) - Transfers connection make direction optional
+- [5762](https://github.com/vegaprotocol/vega/issues/5762) - Transfers connection add under `party` type
 
 ### 🐛 Fixes
 - [5571](https://github.com/vegaprotocol/vega/issues/5571) - Restore pending assets status correctly after snapshot restore
@@ -61,6 +64,7 @@ The init command now also generate the configuration for tendermint, the flags `
 - [5781](https://github.com/vegaprotocol/vega/issues/5781) - Make enactment timestamp optional in proposal for `graphql`
 - [5767](https://github.com/vegaprotocol/vega/issues/5767) - Fix typo in command validation
 - [5825](https://github.com/vegaprotocol/vega/issues/5825) - Fix panic in pegged orders when going into auction
+- [5763](https://github.com/vegaprotocol/vega/issues/5763) - Transfers connection rename `pubkey` to `partyId`
 
 ## 0.53.0